+++ teensy/src/sfr.hpp
@@ -10,7 +10,7 @@ namespace sfr
     namespace anemometer
     {
         extern uint16_t wind_angle;
-        extern float wind_speed;
+        extern uint16_t wind_speed;
     }
     namespace servo
     {
